Rotation of the Stage
Update:
- Rotate stage pivot

ToDo next:
- Fix bugs in the stage position
- Setup player's physics
- Centralize Control Inputs

Notes:
Move a stage pivot and and them correct the stage position is scaling up in complexity, and it feels a bit unecessary. I will try another aproach using just the camera movment and the player. If it not works I will come back to this branch.